+++ libs/functional/tests/unit/function/function_arith.cpp
@@ -14,33 +14,33 @@
 #include <hpx/functional/function.hpp>
 #include <hpx/testing.hpp>
 
-float mul_ints(int x, int y)
+double mul_ints(int x, int y)
 {
-    return ((float) x) * y;
+    return ((double) x) * y;
 }
 
 struct int_div
 {
-    float operator()(int x, int y) const
+    double operator()(int x, int y) const
     {
-        return ((float) x) / y;
+        return ((double) x) / y;
     };
 };
 
 int main()
 {
-    hpx::util::function_nonser<float(int x, int y)> f;
+    hpx::util::function_nonser<double(int x, int y)> f;
 
     f = int_div();
     HPX_TEST(f);
-    HPX_TEST_EQ(f(5, 3), 5.f / 3);
+    HPX_TEST_EQ(f(5, 3), 5. / 3);
 
     f = nullptr;
     HPX_TEST(!f);
 
     f = &mul_ints;
     HPX_TEST(f);
-    HPX_TEST_EQ(f(5, 3), 5.f * 3);
+    HPX_TEST_EQ(f(5, 3), 5. * 3);
 
     return hpx::util::report_errors();
 }
